Refatorando os métodos do CoreData de AppDelegate Criando um singleton próprio, o DatabaseManager, para os métodos CoreData
