+++ bitcoin-converter/Model/CoinManager.swift
@@ -24,8 +24,7 @@ class CoinManager {
     
     //Acesso ao contexto
     var context: NSManagedObjectContext {
-        let appDelegate = UIApplication.shared.delegate as! AppDelegate
-        return appDelegate.persistentContainer.viewContext
+        return DatabaseManager.shared.persistentContainer.viewContext
     }
     
     let currentArray = ["USD", "AUD", "BRL", "CAD", "CHF", "CLP", "CNY", "DKK", "EUR", "GBP", "HKD", "INR", "ISK", "JPY", "KRW", "NZD", "PLN", "RUB", "SEK", "SGD", "THB", "TRY", "TWD"]
